# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/social-links": {
      "get": {
        "operationId": "SocialLinksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      },
      "post": {
        "operationId": "SocialLinksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSocialLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      }
    },
    "/admin/social-links/reorder": {
      "patch": {
        "operationId": "SocialLinksController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      }
    },
    "/admin/social-links/{id}": {
      "get": {
        "operationId": "SocialLinksController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      },
      "patch": {
        "operationId": "SocialLinksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSocialLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      },
      "delete": {
        "operationId": "SocialLinksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "tags": [
          "admin/social-links"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "admin": {
        "type": "apiKey",
        "in": "header",
        "name": "x-authorization",
        "description": "Bearer <access token>"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "ReorderDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSocialLinkDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "format": "uri"
          },
          "visible": {
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "url"
        ]
      },
      "UpdateSocialLinkDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 60
          },
          "url": {
            "type": "string",
            "maxLength": 500,
            "format": "uri"
          },
          "visible": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}